tena

+++ MyApp/MyApp/MyApp/MVVM/Models/Bill.cs
@@ -10,7 +10,7 @@ namespace MyApp.MVVM.Models
         public int clientId { get; set; }
         public string client { get; set; }
         public double originalAmount { get; set; }
-        public int percentage { get; set; }
+        public double percentage { get; set; }
         public double totalAmount { get; set; }
         public double amountPaid { get; set; }
         public double restAmount { get; set; }
